Code clean-up
Improve includes
Add some documentation
Improve some variable names
Fix two core dump
Remove useless files
