+++ moteur/bullet.h
@@ -3,6 +3,8 @@
 	#include "entity.h"
 #endif
 
+#include "../GfxLib/BmpLib.h"
+
 typedef struct bullet {
 	Position pos;
 	Speed speed;
@@ -10,3 +12,13 @@ typedef struct bullet {
 	DonneesImageRGB* image;
 	struct bullet *nextBullet;
 } Bullet; 
+
+Bullet* createBullet(int x, int y, int dir);
+
+Bullet* getLastBullet(Bullet* list);
+
+void insertQueueBullet(Bullet** list, Bullet* maillon);
+
+void removeBullet(Bullet** list, Bullet** maillon);
+
+Bullet* bulletBeforeOf(Bullet** list, Bullet* maillon);
