+++ moteur/displayGame.c
@@ -1,20 +1,12 @@
-#ifndef GFXLIB_H
-	#include "../GfxLib/GfxLib.h"
-	#define GFXLIB_H
-#endif
-
-#ifndef GAME_H
-	#define GAME_H
-	#include "game.h"
-#endif
-
 #include <stdio.h>
+#include "../GfxLib/GfxLib.h"
 #include "../GfxLib/BmpLib.h"
+#include "game.h"
 #include "background.h"
 
 void displayGame(void){ 
 	background();
-	Game *game = gameEvent();
+	Game *game = getGame();
 	ecrisImage(game->spaceship.pos.x, game->spaceship.pos.y, game->spaceship.image->largeurImage, game->spaceship.image->hauteurImage, game->spaceship.image->donneesRGB);
 	
 	Enemy* bufferEnemy = game->enemies;
@@ -34,7 +26,7 @@ void displayGame(void){
 		ecrisImage(bufferBonus->pos.x, bufferBonus->pos.y, bufferBonus->image->largeurImage, bufferBonus->image->hauteurImage, bufferBonus->image->donneesRGB);
 		bufferBonus = bufferBonus->nextBonus;
 	}
-
+	
 	//Display score
 	char scoreStr[50];
 	sprintf(scoreStr, "Score: %d", game->score);
@@ -45,7 +37,7 @@ void displayGame(void){
 	sprintf(levelStr, "Level: %d", game->level);
 	afficheChaine(levelStr, 20, 5, hauteurFenetre() - 55);
 	
-	//Display life
+	//Display life bar
 	couleurCourante(255, 255, 255);
 	rectangle(2, 2, largeurFenetre()/3.0 + 1, 17);
 	couleurCourante(250, 10, 10);
