+++ moteur/enemy.h
@@ -13,3 +13,13 @@ typedef struct enemy {
 	DonneesImageRGB* image;
 	struct enemy *nextEnemy; 
 } Enemy;
+
+Enemy* createEnemy(int x, int y);
+
+Enemy* getLastEnemy(Enemy* list);
+
+void insertQueueEnemy(Enemy** list, Enemy* maillon);
+
+void removeEnemy(Enemy** list, Enemy** maillon);
+
+Enemy* enemyBeforeOf(Enemy** list, Enemy* maillon);
