+++ moteur/game.h
@@ -1,10 +1,7 @@
 #include "spaceship.h"
 #include "enemy.h"
 #include "bullet.h"
-#ifndef BONUS_H
-	#define BONUS_H
-	#include "bonus.h"
-#endif
+#include "bonus.h"
 
 typedef struct {
 	Spaceship spaceship;
@@ -18,6 +15,8 @@ typedef struct {
 
 void startGame(int begin);
 
+Game* initGame(void);
+
 void nextLevel();
 
 int gameState(void);
@@ -26,30 +25,6 @@ void gamePause(int pause);
 
 void endGame(void);
 
-Game *gameEvent(void);
-
-Game initGame(void);
-
-Enemy* createEnemy();
-
-Bullet* createBullet(int x, int y, int dir);
-
-Bullet* getLastBullet(Bullet* list);
-
-Enemy* getLastEnemy(Enemy* list);
-
-void insertQueueEnemy(Enemy** list, Enemy* maillon);
-
-void insertQueueBullet(Bullet** list, Bullet* maillon);
-
-void removeBullet(Bullet** list, Bullet** maillon);
-
-void removeEnemy(Enemy** list, Enemy** maillon);
-
-Bullet* bulletBeforeOf(Bullet** list, Bullet* maillon);
-
-Enemy* enemyBeforeOf(Enemy** list, Enemy* maillon);
-
-int getRand(int max);
+Game* getGame(void);
 
 void activateBonus(Game* game, Bonus perk);
