+++ moteur/param.c
@@ -1,5 +1,5 @@
-#include <stdio.h>
 #include <stdlib.h>
+#include <stdio.h>
 #include "../Sauvegarde/sauvegarde.h"
 
 typedef enum {Defaut, Terminator, MichelATord, FoconMillenium}
