+++ moteur/utils.c
@@ -1,0 +1,6 @@
+#include <stdlib.h>
+
+// Return a random number between 0 and max (exclusive)
+int getRand(int max) {
+	return rand()%max;
+}
